обработка информации о статьях вынесена в класс ArticlesInfo
